Add proper quoting of object names.
The schema and table names in the code are not properly quoted which can cause
problems when deploying scripts containing object names with spaces and other
reserved characters in SQL Server. A method has been added which wraps the object
names in square brackets and escapes the special characters properly where needed.