Refactored StoreExecutedScript for extensibility
I found myself wanting to add extra columns to the SchemaVersions table to store additional information - fortunately there were extensibility points present in this class to add additional columns to the table. However, when I came to insert into it, the StoreExecutedScript method was not virtual and the insert script and db parameters were embedded inside of this method.

All I've done is simply move them out and make them 'protected virtual' to allow for inserting into columns that others may want to add.