Fix nested multiline comment handling by SqlParser
When the SqlParser encountered a nested multiline comment that had the
closing comment delimeters together in sequence it would fail to
properly parse.

Given the example:

    /*ab/*cd*/*/

The SqlParser correctly identifies the second multiline comment
(i.e. "cd") it then advances until the "/" character which closes the
inner comment (10th character in the string). At this character, the
parser properties were looking forward and seeing "/*" but also looking
backwards and seeing "*/". Because the check for the start of another
nested comment was checked first it would be read as the start of
another multiline comment.