Additional FileSystemScriptProvider Documentation
There was a request made about how to configure DbUp to include only
certain files from the file system. This information was missing from
the documentation so that situation is resolved here.

All variants of the ".WithScriptsFromFileSystem(...)" configuration
extension method are here except for the ones which take a
"SqlScriptOptions" type. That type requires additional documentation and
there didn't seem to be a straightforward entry point into that
knowledge.

Related to #501