Do Not Build on Push
Having both "push" and "pull_request" meant that mulitple builds were
being done for one change to a PR. A branch can be built if necessary by
triggering the manual "workflow_dispatch" task. Unless this is
specifically necessary, though, the work will still be built when a PR
is created.